Add DeSelect func to SelectMenu & affected components (#198)
* Add DeSelect func to SelectMenu & affected components

Follow up from issue #196

* Add Multi-Select with Deselect Example to Storybook & Docs

Added onDeselect to index.stories.js, and attempted to put the proper docs files in place. Let me know of changes/corrections and I am happy to fix!

* Fix multi example

Linting helped me reformat this, forgot to update prior to commit..

* Fix multi deselect per linting

* Updates per comments

I also added the if statement (in lieu of switch) solution to the onSelect within these changes. This way it is always number of items + "..." for the Button presentation.

Co-Authored-By: jeroenransijn <jssrdesign@gmail.com>

* Remove switch comments (fix local diff issue)

Something wonky happened with the disk version & VS Code version diff.. not sure, but removing commented switch stmnt